% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{b.cor}
\alias{b.cor}
\title{A function to compute the Covariance}
\usage{
b.cor(x, i)
}
\arguments{
\item{x}{the samples}

\item{i}{the row of the samples}
}
\value{
the covariance of the samples
}
\description{
A function to compute the Covariance
}
