% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{BA}
\alias{BA}
\title{A function to generate BA models}
\usage{
BA(m0, m, N)
}
\arguments{
\item{m0}{Number of nodes in the initial network}

\item{m}{The number of connected edges of newly added nodes at each moment}

\item{N}{sizes of the network}
}
\value{
a BA model network
}
\description{
A function to generate BA models
}
\examples{
\dontrun{
BA(10,2,10000)
}
}
