% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{count5test}
\alias{count5test}
\title{A function to do Count Five Test}
\usage{
count5test(x, y)
}
\arguments{
\item{x}{samples}

\item{y}{paired samples}
}
\value{
return 1 (reject) or 0 (do not reject H0)
}
\description{
A function to do Count Five Test
}
\examples{
{
x=rnorm(100,0,1)
y=rnorm(100,0,1)
count5test(x,y)
}
}
