% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{eqdist.nn}
\alias{eqdist.nn}
\title{A function to compute the results of the NN methods}
\usage{
eqdist.nn(z, sizes, k)
}
\arguments{
\item{z}{data}

\item{sizes}{the sizes of the samples}

\item{k}{The maximum number of nearest neighbours to compute}
}
\value{
the results of the NN methods
}
\description{
A function to compute the results of the NN methods
}
