% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{Ftest}
\alias{Ftest}
\title{A function to do F Test}
\usage{
Ftest(x, y, n)
}
\arguments{
\item{x}{samples}

\item{y}{paired samples}

\item{n}{size of the sample}
}
\value{
return 1 (reject) or 0 (do not reject H0)
}
\description{
A function to do F Test
}
\examples{
{
x=rnorm(100,0,1)
y=rnorm(100,0,1)
Ftest(x,y,100)
}
}
