% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{Gelman.Rubin}
\alias{Gelman.Rubin}
\title{A function using the Gelman-Rubin Method}
\usage{
Gelman.Rubin(psi)
}
\arguments{
\item{psi}{the statistic psi}
}
\value{
the G-R statistc
}
\description{
A function using the Gelman-Rubin Method
}
