% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{HV}
\alias{HV}
\title{A function to generate hidden variable network models}
\usage{
HV(N, beta)
}
\arguments{
\item{N}{sizes of the network}

\item{beta}{a param in the connection function}
}
\value{
a hidden variable network model
}
\description{
A function to generate hidden variable network models
}
\examples{
\dontrun{
HV(10000,0.7)
}
}
