% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{laplace.chain}
\alias{laplace.chain}
\title{A function using to generate the M-H sampler}
\usage{
laplace.chain(sigma, N, X1)
}
\arguments{
\item{sigma}{Variances  used for the proposal distribution}

\item{N}{the number of samples}

\item{X1}{Initial value}
}
\value{
a random sample of size \code{n}
}
\description{
A function using to generate the M-H sampler
}
