% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{maxout}
\alias{maxout}
\title{A function  counts the maximum number of extreme points of each sample with respect to the range of the other sample}
\usage{
maxout(x, y)
}
\arguments{
\item{x}{one sample}

\item{y}{the other sample}
}
\value{
the maximum number of extreme points of each sample with respect to the range of the other sample
}
\description{
A function  counts the maximum number of extreme points of each sample with respect to the range of the other sample
}
\examples{
{
x=1:10
y=2:11
maxout(x,y)
}
}
