% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{mvrnorm}
\alias{mvrnorm}
\title{A function to generate a sample with Multivariate normal distribution}
\description{
A function to generate a sample with Multivariate normal distribution
}
\examples{
{
sigma1=matrix(c(1,0,0,1),2,2)
MASS::mvrnorm(1,rep(0,2),sigma1)
}
}
