% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMC}
\alias{rwMC}
\title{A Metropolis sampler using Rcpp}
\usage{
rwMC(sigma, x0, N)
}
\arguments{
\item{sigma}{Variances  used for the proposal distribution}

\item{x0}{Initial value}

\item{N}{the number of samples}
}
\value{
a random sample of size \code{n} and the rejection rate
}
\description{
A Metropolis sampler using Rcpp
}
\examples{
\dontrun{
rw=rwMC(2,25,1000)
plot(rw$x,type="l",lwd=2,xlab="sigma=2",ylab="X")
}
}
