% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{select}
\alias{select}
\title{A function using to generate the rescaled Epanechnikov kernel sampler}
\usage{
select(n)
}
\arguments{
\item{n}{the number of samples}
}
\value{
a random sample
}
\description{
A function using to generate the rescaled Epanechnikov kernel sampler
}
\examples{
{
y=numeric(10000)
for(i in 1:10000){
y[i]=select(3)
}
print(y[1:10])
}
}
