% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{sk}
\alias{sk}
\title{A function to compute the skewness}
\usage{
sk(x)
}
\arguments{
\item{x}{the samples}
}
\value{
the skewness of the samples
}
\description{
A function to compute the skewness
}
\examples{
{
x=c(1,3,6,10)
sk(x)
}
}
