% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{skm}
\alias{skm}
\title{A function to compute the skewness of a  multivariant sample}
\usage{
skm(x, sigma, n)
}
\arguments{
\item{x}{the samples}

\item{sigma}{the variance of the distribution}

\item{n}{size of the sample}
}
\value{
the skewness of the samples
}
\description{
A function to compute the skewness of a  multivariant sample
}
