\name{StatComp20038-package}
\alias{StatComp20038-package}
\alias{StatComp20038}
\docType{package}
\title{
  This package provides several functions for 'Statistical Computing' course 
}
\description{
  This package provides two functions for generating scale-free network and all functions for my homework.
}
\author{
Zi'ang Xu

Maintainer: Zi'ang Xu <3127224710@qq.com>
}
\keyword{ package }
