% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{Tn}
\alias{Tn}
\title{A function to compute boot statistic}
\usage{
Tn(z, ix, sizes, k)
}
\arguments{
\item{z}{data}

\item{ix}{the ixth data}

\item{sizes}{the sizes of the samples}

\item{k}{The maximum number of nearest neighbours to compute}
}
\value{
boot statistic
}
\description{
A function to compute boot statistic
}
