% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{Ttest}
\alias{Ttest}
\title{A function to do Paired-t Test}
\usage{
Ttest(x, n)
}
\arguments{
\item{x}{samples}

\item{n}{size of the sample}
}
\value{
return 1 (reject) or 0 (do not reject H0)
}
\description{
A function to do Paired-t Test
}
\examples{
{
x=rnorm(100,0,1)
y=rnorm(100,0,1)
Ttest(x-y,100)
}
}
